Errors solved
The code now works I tried a lot of different code to try and make this work and in the end I added.

the method pokemonIndex that all pokemon have and changes depending on where in the list pokemons it is.

the method SetPokemonIndex which sets the pokemon index to what index it has in the list. Everytime a pokemon is added or removed this method is called on so that the index gets updated.

I also added two Hash set strings called pokemonsRemoved and pokemonsAdded, that keeps track so that you can't have 2 of the same pokemon and you can't throw a pokemon out twice.