Locations
I have added the class locations, each location have different actions you can do. I reconnected so that trainer runs the action methods and when it affects the collection the location connects to the collection so it goes from program to Trainer to Location to collection. I haven't fixed completely so that you can add pokemons to inventory yet and I commented the code that removes them because it's too much to figure out right now.